{"schema":"decodex/commit/1","summary":"Move scroll capture image stacking","authority":"manual"}
